requirements.txt
sounddevice
openai
pydub
python-dotenv
keyboard
wave
soundfile
clipboard
numpy